src/centralserver/admin-service/src/main/java/org/niis/xroad/centralserver/restapi/openapi/InitializationApiController.java
package org.niis.xroad.centralserver.restapi.openapi;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.niis.xroad.centralserver.openapi.InitializationApi;
import org.niis.xroad.centralserver.openapi.model.InitialServerConf;
import org.niis.xroad.centralserver.openapi.model.InitializationStatus;
import org.niis.xroad.restapi.openapi.ControllerUtil;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;

@Controller
@RequestMapping(ControllerUtil.API_V1_PREFIX)
@Slf4j
@PreAuthorize("denyAll")
@RequiredArgsConstructor
public class InitializationApiController implements InitializationApi {

    @Override
    @PreAuthorize("isAuthenticated()")
    public ResponseEntity<InitializationStatus> getInitializationStatus() {
        return null;
    }

    @Override
    @PreAuthorize("hasAuthority('INIT_CONFIG')")
    public ResponseEntity<Void> initCentralServer(InitialServerConf initialServerConf) {
        return null;
    }
}
